**Before**

```
➜  rector-src git:(main) time bin/rector process --clear-cache
 2269/2269 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        
bin/rector process --clear-cache  243.43s user 13.41s system 700% cpu 36.678 total
```

**After**

```
➜  rector-src git:(performance-early-already-last-name-usage) time bin/rector process --clear-cache                 
 2269/2269 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        
bin/rector process --clear-cache  223.70s user 10.11s system 734% cpu 31.843 total
```