@samsonasik 2024-08-29
Tested in CodeIgniter 4 project, 3 seconds faster:

**Before**

```
➜  CodeIgniter4 git:(develop) time vendor/bin/rector process --clear-cache
 956/956 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
      
vendor/bin/rector process --clear-cache  317.14s user 8.81s system 713% cpu 45.694 total
```

**After**

```
➜  CodeIgniter4 git:(develop) time vendor/bin/rector process --clear-cache
 956/956 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        
vendor/bin/rector process --clear-cache  296.78s user 7.59s system 716% cpu 42.456 total
```