@samsonasik 2024-08-29
Tested in a project with **4869 files**, got **87 seconds** faster, a clear win 🎉 🎉 🎉 

**Before**

```
vendor/bin/rector process --clear-cache  1452.18s user 38.23s system 624% cpu 3:58.81 total
```

**After**

```
vendor/bin/rector process --clear-cache  1084.33s user 37.76s system 741% cpu 2:31.29 total
```